1.21.1 API fixes round 2: InteractionResultHolder use(), AQUA, BLOCKS, 3-arg getHeight, unbounded nextFloat, Vec3 multiply, BlockPos casts, sendCustomPayload; decisive API probes round 2

+++ src/main/java/com/tornadic/tornado/TornadoState.java
@@ -71,7 +71,7 @@ public class TornadoState {
 	}
 
 	public float funnelRadius() {
-		float base = TornadoIntensity.fromEf(peakEf).funnelRadius();
+		float base = TornadoIntensity.fromEf((int) peakEf).funnelRadius();
 		float progress = Math.max(0.25F, intensity);
 		return base * (0.4F + 0.6F * progress);
 	}
